Superhirn PATCH by Florian Zeller: Fixed conversion double->float problem (compiler warning was right)
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@10070 6adb6e08-d915-0410-941f-83917bcadc18

+++ include/OpenMS/TRANSFORMATIONS/FEATUREFINDER/SUPERHIRN/LC_elution_peak.h
@@ -249,7 +249,7 @@ public:
   ///////////////////
   // get scan apex:
   int get_scan_apex(){return fScanNumberApex;};
-  float get_apex_intensity(){return fapex_intensity;};
+  double get_apex_intensity(){return fapex_intensity;};
   double get_apex_retention_time(){return fRT;};
   double get_apex_MZ(){return get_MZ(get_scan_apex());};
   
@@ -261,7 +261,7 @@ public:
   
   /////////////////
   // get the total peak area:
-  float get_total_peak_area(){return fpeak_area;};
+  double get_total_peak_area(){return fpeak_area;};
   
   ////////////////
   // get start / end scan:
